+++ src/elixir/Enum/index.ts
@@ -1,4 +1,5 @@
 export { chunk_every } from './chunk_every.ts'
 export { frequencies_by } from './frequencies_by.ts'
 export { group_by } from './group_by.ts'
+export { reduce_while } from './reduce_while.ts'
 export { scan } from './scan.ts'