+++ src/python/index.ts
@@ -1,3 +1,4 @@
+export * as difflib from './difflib/index.ts'
 export * as math from './math/index.ts'
 export * as re from './re/index.ts'
 export * as string from './string/index.ts'